backend/requirements.txt
openpyxl~=3.1.0
gspread
googleapiclient
python-dotenv
Flask
